Creacionde DAO Entity, Crosscuting y catalogo de Mensajes

+++ src/main/java/co/edu/uco/spaonline/dto/ServicioDTO.java
@@ -1,6 +1,6 @@
 package co.edu.uco.spaonline.dto;
 
-import java.math.BigDecimal;
+
 import java.util.UUID;
 
 public final class ServicioDTO {
@@ -9,13 +9,13 @@ public final class ServicioDTO {
 	private String nombre;
 	private String descipcion;
 	private String tiposervicio;
-	private BigDecimal tarifa;
+	private Long tarifa;
 	
 	private ServicioDTO() {
 		super();
 	}
 
-	private ServicioDTO(final UUID id,final String nombre,final String descipcion,final String tiposervicio,final BigDecimal tarifa) {
+	private ServicioDTO(final UUID id,final String nombre,final String descipcion,final String tiposervicio,final Long tarifa) {
 		setId(id);
 		setNombre(nombre);
 		setDescipcion(descipcion);
@@ -63,11 +63,11 @@ public final class ServicioDTO {
 		return this;
 	}
 	
-	public final BigDecimal getTarifa() {
+	public final Long getTarifa() {
 		return tarifa;
 	}
 	
-	public final ServicioDTO setTarifa(BigDecimal tarifa) {
+	public final ServicioDTO setTarifa(Long tarifa) {
 		this.tarifa = tarifa;
 		return this;
 	}
